separate building of main, test and benchmark
