+++ CMakeLists.txt
@@ -43,6 +43,8 @@ endif()
 if(${ENABLE_UBSAN})
     set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -fsanitize=undefined -fno-common -fpermissive")
 endif()
+option(BUILD_TESTS "Build C++ and Python tests." FALSE)
+option(BUILD_BENCHMARK "Build benchmarks." FALSE)
 
 function(link_threads LIBRARY)
     if (CMAKE_VERSION VERSION_LESS "3.1")
@@ -55,7 +57,7 @@ endfunction()
 function(add_kuzu_test TEST_NAME)
     set(SRCS ${ARGN})
     add_executable(${TEST_NAME} ${SRCS})
-    target_link_libraries(${TEST_NAME} PRIVATE test_helper)
+    target_link_libraries(${TEST_NAME} PRIVATE test_helper graph_test)
     target_include_directories(${TEST_NAME} PRIVATE ${PROJECT_SOURCE_DIR}/test/include)
     include(GoogleTest)
     gtest_discover_tests(${TEST_NAME})
@@ -82,5 +84,9 @@ include_directories(third_party/pybind11/include)
 
 add_subdirectory(third_party)
 add_subdirectory(src)
-add_subdirectory(test)
+if (${BUILD_TESTS})
+    add_subdirectory(test)
+elseif (${BUILD_BENCHMARK})
+    add_subdirectory(test/test_helper)
+endif()
 add_subdirectory(tools)
