+++ test/catalog/CMakeLists.txt
@@ -1,5 +1,5 @@
 add_executable(catalog_test catalog_test.cpp)
 
-target_link_libraries(catalog_test PUBLIC test_helper)
+target_link_libraries(catalog_test PUBLIC graph_test)
 
 add_test(kuzu_catalog_test catalog_test)
