+++ test/common/time_test.cpp
@@ -2,10 +2,9 @@
 
 #include "common/exception.h"
 #include "common/types/types_include.h"
-#include "include/gtest/gtest.h"
+#include "gtest/gtest.h"
 
 using namespace kuzu::common;
-using namespace std;
 
 TEST(TimeTests, FromTime) {
     // Hour out of range
