+++ test/copy/copy_fault_test.cpp
@@ -1,8 +1,7 @@
 #include "common/exception.h"
-#include "test_helper/test_helper.h"
+#include "graph_test/graph_test.h"
 
 using namespace kuzu::testing;
-using namespace std;
 
 class CopyFaultTest : public EmptyDBTest {
 public:
