+++ test/copy/copy_interval_test.cpp
@@ -1,8 +1,7 @@
 #include <string>
 
-#include "test_helper/test_helper.h"
+#include "graph_test/graph_test.h"
 
-using namespace std;
 using namespace kuzu::common;
 using namespace kuzu::storage;
 using namespace kuzu::testing;
