+++ test/copy/copy_lists_test.cpp
@@ -1,7 +1,6 @@
 #include "common/type_utils.h"
-#include "test_helper/test_helper.h"
+#include "graph_test/graph_test.h"
 
-using namespace std;
 using namespace kuzu::common;
 using namespace kuzu::storage;
 using namespace kuzu::testing;
