+++ test/copy/copy_test.cpp
@@ -1,7 +1,6 @@
 #include "common/csv_reader/csv_reader.h"
-#include "test_helper/test_helper.h"
+#include "graph_test/graph_test.h"
 
-using namespace std;
 using namespace kuzu::common;
 using namespace kuzu::catalog;
 using namespace kuzu::storage;
