+++ test/include/main_test_helper/main_test_helper.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "test_helper/test_helper.h"
+#include "graph_test/graph_test.h"
 
 namespace kuzu {
 namespace testing {
