+++ test/runner/e2e_ddl_test.cpp
@@ -1,5 +1,5 @@
+#include "graph_test/graph_test.h"
 #include "processor/mapper/plan_mapper.h"
-#include "test_helper/test_helper.h"
 
 using namespace kuzu::testing;
