+++ test/runner/e2e_delete_create_transaction_test.cpp
@@ -1,6 +1,4 @@
-#include <iostream>
-
-#include "test_helper/test_helper.h"
+#include "graph_test/graph_test.h"
 
 using namespace kuzu::testing;
