+++ test/transaction/transaction_test.cpp
@@ -1,5 +1,5 @@
 #include "common/configs.h"
-#include "test_helper/test_helper.h"
+#include "graph_test/graph_test.h"
 
 using namespace kuzu::testing;
